@@ -1,6 +1,6 @@
 {
   "name": "cybex-web",
-  "version": "1.1.0",
+  "version": "1.2.0",
   "description": "An Advanced wallet interface, inspired by bitshares-ui",
   "author": "CybexDEX",
   "browser": {
@@ -122,8 +122,9 @@
     "@fortawesome/fontawesome-free-brands": "^5.0.6",
     "@fortawesome/fontawesome-free-solid": "^5.0.6",
     "@fortawesome/react-fontawesome": "^0.0.17",
+    "@types/d3-dsv": "^1.0.31",
     "@types/prop-types": "^15.5.2",
-    "@types/react": "15",
+    "@types/react": "^16.3.13",
     "@types/react-svg": "^2.2.0",
     "ReconnectingWebSocket": "git+https://github.com/bitshares/reconnecting-websocket",
     "alt": "^0.18.6",
@@ -137,13 +138,18 @@
     "bufferutil": "^3.0.2",
     "bytebuffer": "^5.0.0",
     "chart.js": "^2.7.0",
+    "chroma-js": "^1.3.7",
     "classnames": "^2.2.1",
     "cli-usage": "^0.1.4",
     "cookies-js": "^1.2.1",
     "counterpart": "^0.18.3",
     "create-hash": "^1.1.2",
     "create-hmac": "^1.1.4",
+    "create-react-class": "^15.6.3",
     "crypto-js": "^3.1.9-1",
+    "d3-dsv": "^1.0.8",
+    "d3-format": "^1.2.2",
+    "d3-interpolate": "^1.1.6",
     "deep-equal": "^1.0.1",
     "ecurve": "^1.0.2",
     "event-emitter": "^0.3.4",
@@ -169,28 +175,31 @@
     "perfect-scrollbar": "git+https://github.com/bitshares/perfect-scrollbar.git",
     "preload-webpack-plugin": "^2.0.0",
     "prop-types": "^15.6.1",
+    "pubsub-js": "^1.6.0",
     "qrcode.react": "^0.7.1",
+    "radium": "^0.24.0",
     "rc-queue-anim": "^1.2.2",
-    "react": "^15.6.1",
+    "react": "^16.3.2",
     "react-apollo": "^2.0.4",
     "react-clipboard.js": "^1.0.1",
-    "react-dom": "^15.6.1",
-    "react-foundation-apps": "git+https://github.com/valzav/react-foundation-apps.git",
+    "react-dom": "^16.3.2",
     "react-highcharts": "^12.0.0",
     "react-hot-loader": "^3.0.0",
     "react-interpolate-component": "^0.10.0",
     "react-intl": "^2.4.0",
     "react-json-inspector": "^7.0.0",
-    "react-notification-system": "^0.2.16",
-    "react-popover": "^0.4.6",
+    "react-notification-system": "^0.2.17",
+    "react-popover": "^0.5.5",
     "react-responsive-mixin": "^0.4.0",
     "react-router": "^3.0.2",
-    "react-stockcharts": "^0.6.1",
+    "react-select": "^2.0.0-beta.4",
+    "react-stockcharts": "file:///Users/cy-work/projects/react-stockcharts/build",
     "react-svg": "^2.2.18",
     "react-svg-inline": "^2.1.0",
     "react-tooltip": "^3.2.2",
     "react-transition-group": "^1.2.0",
-    "react-translate-component": "^0.13.2",
+    "react-translate-component": "^0.15.1",
+    "redux": "^4.0.0",
     "rxjs": "^5.4.3",
     "secure-random": "^1.1.1",
     "tcomb": "2.5.2",
@@ -200,8 +209,13 @@
     "zxcvbn": "^4.4.2"
   },
   "devDependencies": {
+    "@types/d3-format": "^1.2.2",
+    "@types/d3-interpolate": "^1.1.6",
+    "@types/d3-time": "^1.0.7",
+    "@types/d3-time-format": "^2.1.0",
     "@types/lodash": "^4.14.92",
-    "@types/react-dom": "15",
+    "@types/radium": "^0.24.0",
+    "@types/react-dom": "^16.0.5",
     "@types/react-router": "3",
     "awesome-typescript-loader": "^3.2.3",
     "babel-core": "^6.21.0",
@@ -220,6 +234,9 @@
     "copyfiles": "^1.2.0",
     "cross-env": "^5.0.5",
     "css-loader": "^0.22.0",
+    "d3-scale": "1",
+    "d3-time": "^1.0.8",
+    "d3-time-format": "^2.1.1",
     "electron": "^1.7.9",
     "electron-builder": "19.43.0",
     "es6-promise": "^3.0.2",
@@ -247,7 +264,7 @@
     "style-loader": "^0.13.1",
     "svg-inline-loader": "^0.7.1",
     "svgo-loader": "1.1.2",
-    "typescript": "^2.5.2",
+    "typescript": "^2.8.3",
     "uglify-js": "^3.2.1",
     "uglifyjs-webpack-plugin": "^1.1.2",
     "url-loader": "^0.5.7",
