@@ -1,9 +1,15 @@
-if (!window.Intl) { // Safari polyfill
-    require.ensure(["intl"], require => {
-        window.Intl = require("intl");
-        Intl.__addLocaleData(require("./assets/intl-data/en.json"));
-        require("index.js");
-    });
-} else {
+import createClass from "create-react-class";
+import * as PropTypes from "prop-types";
+import * as React from "react";
+React.createClass = createClass;
+React.PropTypes = PropTypes;
+if (!window.Intl) {
+  // Safari polyfill
+  require.ensure(["intl"], require => {
+    window.Intl = require("intl");
+    Intl.__addLocaleData(require("./assets/intl-data/en.json"));
     require("index.js");
+  });
+} else {
+  require("index.js");
 }
