@@ -7,9 +7,9 @@ module.exports = {
         callColor: "#BBBF2B",
         settleColor: "rgba(125, 134, 214, 1)",
         settleFillColor: "rgba(125, 134, 214, 0.5)",
-        positiveColor: "#04B528",
+        positiveColor: "#6DBB49",
         // positiveColor: "#6BA583",
-        negativeColor: "#DD0000",
+        negativeColor: "#D24632",
         primaryText: "#e0e0e0",
         volumeColor: "#848484",
 
@@ -22,75 +22,6 @@ module.exports = {
         axisLineColor: "#AAAAAA",
         indicatorLineColor: "#FFFFFF"
 
-    },
-    olDarkTheme: {
-        bidColor: "#50D2C2",
-        bidFillColor: "rgba(80, 210, 194, 0.5)",
-        askColor: "#E3745B",
-        askFillColor: "rgba(227, 116, 91, 0.5)",
-        callColor: "#BBBF2B",
-        settleColor: "rgba(125, 134, 214, 1)",
-        settleFillColor: "rgba(125, 134, 214, 0.5)",
-        positiveColor: "#6BA583",
-        negativeColor: "#DB0000",
-        strokeColor: "#FFFF00",
-        primaryText: "#e0e0e0",
-        volumeColor: "#848484",
-
-         //tooltip
-        tooltipBackgroundColor: "rgba(0, 0, 0, 0.4)",
-        tooltipColor: "#fff",
-        tooltipFillColor: "#fff",
-        //axis
-        axisLabelsColor: "#fff",
-        axisLineColor: "#AAAAAA",
-        indicatorLineColor: "#FFFFFF"
-
-    },
-    darkTheme: {
-        bidColor: "#6BA583",
-        bidFillColor: "rgba(80, 210, 194, 0.5)",
-        askColor: "#E3745B",
-        askFillColor: "rgba(227, 116, 91, 0.5)",
-        callColor: "#BBBF2B",
-        settleColor: "rgba(125, 134, 214, 1)",
-        settleFillColor: "rgba(125, 134, 214, 0.5)",
-        positiveColor: "#258A14",
-        negativeColor: "#DB0000",
-        strokeColor: "#FFFF00",
-        primaryText: "#e0e0e0",
-        volumeColor: "#848484",
-
-        //tooltip
-        tooltipBackgroundColor: "rgba(0, 0, 0, 0.4)",
-        tooltipColor: "#fff",
-        tooltipFillColor: "#fff" ,
-        //axis
-        axisLabelsColor: "#fff",
-        axisLineColor: "#AAAAAA",
-        indicatorLineColor: "#FFFFFF"
-    },
-    lightTheme: {
-        bidColor: "#258A14",
-        bidFillColor: "rgba(80, 210, 194, 0.5)",
-        askColor: "#EA340B",
-        askFillColor: "rgba(227, 116, 91, 0.5)",
-        callColor: "#BBBF2B",
-        settleColor: "rgba(125, 134, 214, 1)",
-        settleFillColor: "rgba(125, 134, 214, 0.5)",
-        positiveColor: "#528c0a",
-        negativeColor: "rgba(225, 66, 74, 1)",
-        strokeColor: "#000000",
-        primaryText: "#242424",
-        volumeColor: "#848484",
-
-        //tooltip
-        tooltipBackgroundColor: "rgba(255,255,255, 0.9)",
-        tooltipColor: "#000",
-        tooltipFillColor: "#000",
-        //axis
-        axisLabelsColor: "#000",
-        axisLineColor: "#000",
-        indicatorLineColor: "#848484"
     }
+
 }
