@@ -75,11 +75,11 @@ export default class ExchangeHeader extends React.Component {
     const volumeQuote = marketStats.get("volumeQuote");
 
     return (
-      <div className="grid-block shrink no-padding overflow-visible top-bar bgcolor-primary with-shadow">
+      <div className="grid-block shrink no-padding overflow-visible top-bar">
         <div className="grid-block overflow-visible">
           <div className="grid-block shrink show-for-large">
-            <a
-              className="v-align"
+            <span
+              className="v-align clickable"
               onClick={this._addMarket.bind(this, quoteSymbol, baseSymbol)}
             >
               <span style={{ paddingRight: 0 }} className="market-symbol">
@@ -92,97 +92,93 @@ export default class ExchangeHeader extends React.Component {
                 {" / "}
                 <AssetName name={baseSymbol} />
               </span>
-            </a>
+            </span>
           </div>
 
           <div className="grid-block vertical" style={{ overflow: "visible" }}>
-            <div className="grid-block wrap market-stats-container">
-              <ul className="market-stats stats top-stats medium-12">
-                {latestPrice ? (
-                  <PriceStat
-                    ready={marketReady}
-                    price={latestPrice.full}
-                    quote={quoteAsset}
-                    base={baseAsset}
-                    hideQuote={true}
-                    content="exchange.latest"
-                  />
-                ) : null}
+            <ul className="market-stats stats top-stats medium-12">
+              {latestPrice ? (
+                <PriceStat
+                  ready={marketReady}
+                  price={latestPrice.full}
+                  quote={quoteAsset}
+                  base={baseAsset}
+                  hideQuote={true}
+                  content="exchange.latest"
+                />
+              ) : null}
 
-                <li className="stat">
-                  <span>
-                    <Translate
-                      className="nowrap"
-                      component="span"
-                      content="account.hour_24"
-                    />
-                    <br />
-                    <b className={"value " + dayChangeClass}>
-                      {marketReady ? dayChange : 0}
-                      <span className={dayChangeArrow}>
-                        &nbsp;{dayChangeArrow === "" ? null : dayChangeArrow ===
-                        "change-up" ? (
-                          <span>&#8593;</span>
-                        ) : (
-                          <span>&#8595;</span>
-                        )}
-                      </span>
-                    </b>
-                    <span>%</span>
-                  </span>
-                </li>
-
-                {volumeBase >= 0 ? (
-                  <PriceStat
-                    ready={marketReady}
-                    decimals={0}
-                    volume={true}
-                    price={volumeBase}
-                    className="column-hide-small"
-                    volume2={volumeQuote}
-                    base={baseAsset}
-                    quote={quoteAsset}
-                    content="exchange.volume_24"
+              <li className="stat">
+                <span>
+                  <Translate
+                    className="nowrap"
+                    component="span"
+                    content="account.hour_24"
                   />
-                ) : null}
+                  <br />
+                  <b className={"value " + dayChangeClass}>
+                    {marketReady ? dayChange : 0}
+                    <span className={dayChangeArrow}>
+                      &nbsp;{dayChangeArrow === "" ? null : dayChangeArrow ===
+                      "change-up" ? (
+                        <span>&#8593;</span>
+                      ) : (
+                        <span>&#8595;</span>
+                      )}
+                    </span>
+                  </b>
+                  <span>%</span>
+                </span>
+              </li>
 
-                {!hasPrediction && feedPrice ? (
-                  <PriceStat
-                    toolTip={counterpart.translate("tooltip.settle_price")}
-                    ready={marketReady}
-                    className="column-hide-small"
-                    price={feedPrice.toReal()}
-                    quote={quoteAsset}
-                    base={baseAsset}
-                    content="exchange.settle"
-                  />
-                ) : null}
+              {volumeBase >= 0 ? (
+                <PriceStat
+                  ready={marketReady}
+                  decimals={0}
+                  volume={true}
+                  price={volumeBase}
+                  volume2={volumeQuote}
+                  base={baseAsset}
+                  quote={quoteAsset}
+                  content="exchange.volume_24"
+                />
+              ) : null}
 
-                {lowestCallPrice && showCallLimit ? (
-                  <PriceStat
-                    toolTip={counterpart.translate("tooltip.call_limit")}
-                    ready={marketReady}
-                    className="column-hide-medium is-call"
-                    price={lowestCallPrice}
-                    quote={quoteAsset}
-                    base={baseAsset}
-                    content="explorer.block.call_limit"
-                  />
-                ) : null}
+              {!hasPrediction && feedPrice ? (
+                <PriceStat
+                  toolTip={counterpart.translate("tooltip.settle_price")}
+                  ready={marketReady}
+                  price={feedPrice.toReal()}
+                  quote={quoteAsset}
+                  base={baseAsset}
+                  content="exchange.settle"
+                />
+              ) : null}
 
-                {feedPrice && showCallLimit ? (
-                  <PriceStat
-                    toolTip={counterpart.translate("tooltip.margin_price")}
-                    ready={marketReady}
-                    className="column-hide-medium is-call"
-                    price={feedPrice.getSqueezePrice({ real: true })}
-                    quote={quoteAsset}
-                    base={baseAsset}
-                    content="exchange.squeeze"
-                  />
-                ) : null}
-              </ul>
-            </div>
+              {lowestCallPrice && showCallLimit ? (
+                <PriceStat
+                  toolTip={counterpart.translate("tooltip.call_limit")}
+                  ready={marketReady}
+                  className="is-call"
+                  price={lowestCallPrice}
+                  quote={quoteAsset}
+                  base={baseAsset}
+                  content="explorer.block.call_limit"
+                />
+              ) : null}
+
+              {feedPrice && showCallLimit ? (
+                <PriceStat
+                  toolTip={counterpart.translate("tooltip.margin_price")}
+                  ready={marketReady}
+                  className="is-call"
+                  price={feedPrice.getSqueezePrice({ real: true })}
+                  quote={quoteAsset}
+                  base={baseAsset}
+                  content="exchange.squeeze"
+                />
+              ) : null}
+            </ul>
           </div>
         </div>
       </div>
