@@ -1,13 +1,55 @@
 // look for more icons here https://linearicons.com/free or here http://hawcons.com/preview/
 
-import * as React from "react"; import * as PropTypes from "prop-types"; 
+import * as React from "react";
+import * as PropTypes from "prop-types";
 
-let icons = ["user", "trash", "chevron-down", "menu", "database", "download", "search",
-    "plus-circle", "question-circle", "cross-circle", "cog", "layers", "users", "wand", "b-logo",
-    "accounts", "witnesses", "assets", "proposals", "blocks", "committee_members", "workers", "key",
-    "checkmark-circle", "checkmark", "piggy", "locked", "unlocked" , "markets", "fi-star" ,"fees",
-    "thumb-tack", "clock", "clippy", "shuffle", "transfer", "dollar", "deposit", "withdraw",
-    "settle", "trade", "adjust", "excel", "share", "minus-circle"];
+let icons = [
+  "user",
+  "trash",
+  "chevron-down",
+  "menu",
+  "database",
+  "download",
+  "search",
+  "plus-circle",
+  "question-circle",
+  "cross-circle",
+  "cog",
+  "layers",
+  "users",
+  "wand",
+  "b-logo",
+  "accounts",
+  "witnesses",
+  "assets",
+  "proposals",
+  "blocks",
+  "committee_members",
+  "workers",
+  "key",
+  "checkmark-circle",
+  "checkmark",
+  "piggy",
+  "locked",
+  "unlocked",
+  "markets",
+  "fi-star",
+  "fees",
+  "thumb-tack",
+  "clock",
+  "clippy",
+  "shuffle",
+  "transfer",
+  "dollar",
+  "deposit",
+  "withdraw",
+  "settle",
+  "trade",
+  "adjust",
+  "excel",
+  "share",
+  "minus-circle"
+];
 
 let icons_map = {};
 for (let i of icons) icons_map[i] = require(`./${i}.svg`);
@@ -15,23 +57,29 @@ for (let i of icons) icons_map[i] = require(`./${i}.svg`);
 require("./icon.scss");
 
 class Icon extends React.Component {
-    render() {
-        let classes = "icon " + this.props.name;
-        if(this.props.size) {
-            classes += " icon-" + this.props.size;
-        }
-        if(this.props.className) {
-            classes += " " + this.props.className;
-        }
-        return <span className={classes} style={this.props.style || {}} dangerouslySetInnerHTML={{__html: icons_map[this.props.name]}}/>;
+  render() {
+    let classes = "icon " + this.props.name;
+    if (this.props.size) {
+      classes += " icon-" + this.props.size;
     }
+    if (this.props.className) {
+      classes += " " + this.props.className;
+    }
+    return (
+      <span
+        className={classes}
+        style={this.props.style || {}}
+        dangerouslySetInnerHTML={{ __html: icons_map[this.props.name] }}
+      />
+    );
+  }
 }
 
 Icon.propTypes = {
-    name: PropTypes.string.isRequired,
-    size: PropTypes.oneOf(["1x", "2x", "3x", "4x", "5x", "10x"]),
-    inverse: PropTypes.bool,
-    className: PropTypes.string
+  name: PropTypes.string.isRequired,
+  size: PropTypes.oneOf(["1x", "2x", "3x", "4x", "5x", "10x"]),
+  inverse: PropTypes.bool,
+  className: PropTypes.string
 };
 
 export default Icon;
