@@ -1,15 +1,15 @@
-import * as React from "react"; import * as PropTypes from "prop-types";
-
+import * as React from "react";
 
 import utils from "common/utils";
 
 class PriceText extends React.Component {
   render() {
-    let { price, preFormattedPrice, quote, base, component } = this.props;
+    let { price, preFormattedPrice, quote, base, component, precision } = this.props;
 
     let formattedPrice = preFormattedPrice
       ? preFormattedPrice
-      : utils.price_to_text(price, quote, base);
+      : utils.price_to_text(price, quote, base, precision);
+    
 
     if (formattedPrice.full >= 1) {
       return (
