@@ -1,4 +1,4 @@
-if (navigator) {
-  navigator.userAgent.search(/windows/i) !== -1 &&
-    document.body.classList.add("scroll-bar-for-win");
-}
+// if (navigator) {
+//   navigator.userAgent.search(/windows/i) !== -1 &&
+//     document.body.classList.add("scroll-bar-for-win");
+// }
